fix(edgy): neo-tree doesn't open when leader isn't space
The default mapping for this is to use `<leader>` which edgy should
respect. This ensures that neo-tree still opens even when leader has
been remapped. Otherwise, the top neo-tree window just stays in a
loading state.

Unfortunately, it seems the `<leader>` isn't a supported keycode for
`nvim_input` (or at least it isn't listed in `:help keycodes` and it
doesn't seem to work). Getting the leader first, then passing it to
`nvim_input` seems to clear this up.