fix(egdy): do not add panes for not enabled neo-tree sources (#3667)
## What is this PR for?

When the user disables a specific neo-tree source (git_status or
buffers), the panes are still created but throw an error when the user
tries to open them

**NB**: This is currently not functional because `LazyVim.opts` is
returning an empty table instead of the neo-tree opts. I'm hoping
someone can give me a hand with this little problem

## Does this PR fix an existing issue?

<!--
  If this PR fixes any issues, please link to the issue here.
  Fixes #<issue_number>
-->

## Checklist

- [x] I've read the
[CONTRIBUTING](https://github.com/LazyVim/LazyVim/blob/main/CONTRIBUTING.md)
guidelines.