feat(bufferline): integration for catppuccin highlights
I'm not sure how to do this if `LazyVim.config.colorscheme` is a
function, because it doesn't return a string to compare to. I also tried
`vim.cmd("colorscheme")`, but it executes too late. If you know of a way
to also include the function please feel free to make any changes
necessary.