docs(README): translate README.md to Spanish (#5054)
## Description

This pull request translates the README.md file from English to Spanish
to provide Spanish-speaking users with an easier way to understand the
setup and features of LazyVim. The translation includes all sections of
the original file, ensuring that the core documentation remains intact
while making it more accessible to a broader audience.

## Related Issue(s)

None.

## Screenshots

None.

## Checklist

- [x] I've read the
[CONTRIBUTING](https://github.com/LazyVim/LazyVim/blob/main/CONTRIBUTING.md)
guidelines.