fix(copilot-chat): map naming discrepancies between modules
This mainly fixes naming discrepancies between LazyVim picker names and
CopilotChat integration names. It should also make this easily
extendable in the future. Also, since it might be possible that the user
completely disables fzf-lua and thus the default fzf picker is not
available, we should capture that rare scenario, which also makes it
future-proof.