fix(lang.python): enable venv-selector even when telescope is not available (#5829)
## Description

Enable the `venv-selector` plugin even when the telescope plugin is not
available. Now the `venv-selector` plugin supports `fzf-lua` and native
selection, and by default, it chooses the previous, if available.

More on this PR:
https://github.com/linux-cultist/venv-selector.nvim/pull/188

Edit: Use the main branch of the repo since it is now the most updated
branch

## Related Issue(s)

None

## Screenshots

N/A

## Checklist

- [x] I've read the
[CONTRIBUTING](https://github.com/LazyVim/LazyVim/blob/main/CONTRIBUTING.md)
guidelines.