fix(extras): adjust switch source/header shortcut (#6567)
The command to switch header and source is now called
`:LspClangdSwitchSourceHeader` and no longer
`:ClangdSwitchSourceHeader`.

## Description

This is just a small fix due to a command that has been renamed.

I did not find the documentation but the relevant line in the code where
the command is defined is here:
-
https://github.com/neovim/nvim-lspconfig/blob/e688b486fe9291f151eae7e5c0b5a5c4ef980847/lsp/clangd.lua#L92

## Related Issue(s)

I did not find any related issues.

## Checklist

- [x] I've read the
[CONTRIBUTING](https://github.com/LazyVim/LazyVim/blob/main/CONTRIBUTING.md)
guidelines.